zoom in zoom out

+++ GameVQSV/src/game/entity/Entity.java
@@ -5,8 +5,8 @@ import java.awt.image.BufferedImage;
 public class Entity {
 	
 	//set player default position
-	public int worldX, worldY;
-	public int speed;
+	public double worldX, worldY;
+	public double speed;
 	
 	//thêm ảnh
 	public BufferedImage up1, up2, down1, down2, left1, left2, right1, right2;
